# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OFFITEC ERP API",
    "version": "1.0.0",
    "description": "OFFITEC Kurumsal ERP Sistemi API Dokümantasyonu"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateCustomerRequest": {
        "type": "object",
        "required": [
          "companyName"
        ],
        "properties": {
          "companyName": {
            "type": "string",
            "example": "Tech Corp A.Ş."
          },
          "segment": {
            "type": "string",
            "example": "Enterprise"
          },
          "taxOffice": {
            "type": "string",
            "example": "Mecidiyeköy"
          },
          "taxNumber": {
            "type": "string",
            "example": "1234567890"
          },
          "address": {
            "type": "string",
            "example": "Levent, İstanbul"
          },
          "mainPhone": {
            "type": "string",
            "example": "+90 212 123 45 67"
          },
          "mainEmail": {
            "type": "string",
            "format": "email",
            "example": "info@techcorp.com"
          }
        }
      },
      "AddCustomerNoteRequest": {
        "type": "object",
        "required": [
          "noteText",
          "noteType"
        ],
        "properties": {
          "noteText": {
            "type": "string",
            "example": "Müşteri ile SLA sözleşmesi yenilendi."
          },
          "noteType": {
            "type": "string",
            "enum": [
              "internal",
              "technical"
            ],
            "example": "internal"
          },
          "isHighlight": {
            "type": "boolean",
            "description": "Kritik not vurgulama",
            "example": true
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/customers": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteri Listesini Getir",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "segment",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Müşteri listesi"
          },
          "401": {
            "description": "Yetkisiz"
          }
        }
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Yeni Müşteri Oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Müşteri başarıyla oluşturuldu"
          },
          "400": {
            "description": "Geçersiz istek"
          }
        }
      }
    },
    "/customers/{id}": {
      "patch": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteri bilgilerini güncelle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyName": {
                    "type": "string"
                  },
                  "segment": {
                    "type": "string"
                  },
                  "taxOffice": {
                    "type": "string"
                  },
                  "taxNumber": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "mainPhone": {
                    "type": "string"
                  },
                  "mainEmail": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Müşteri güncellendi"
          },
          "400": {
            "description": "Geçersiz istek"
          }
        }
      }
    },
    "/customers/{id}/dashboard": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteri 360° Genel Bakış Paneli (Dashboard) Getir",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard verisi başarıyla getirildi"
          },
          "404": {
            "description": "Müşteri bulunamadı"
          }
        }
      }
    },
    "/customers/{id}/notes": {
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteriye Not Ekle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCustomerNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Not başarıyla eklendi"
          }
        }
      }
    },
    "/customers/{id}/activities": {
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteri Zaman Çizelgesine Aktivite Ekle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activityType": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "activityDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Aktivite kaydedildi."
          }
        }
      }
    },
    "/customers/{id}/documents": {
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Müşteriye Ait Doküman Yükle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileName": {
                    "type": "string"
                  },
                  "fileUrl": {
                    "type": "string"
                  },
                  "fileType": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Doküman meta verisi kaydedildi."
          }
        }
      }
    },
    "/customers/{id}/tenders": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Belirli bir müşteriye ait tüm ihaleleri/teklifleri listele",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}